Changes: -TryCatch Blocks -InvoiceService Discount Function Update have been made

+++ ShopsRUs/Dtos/InvoiceDto.cs
@@ -12,7 +12,7 @@ namespace ShopsRUs.Dtos
         public string InvoiceId { get; set; } 
         public string CustomerId { get; set; }
         public double AllowanceAmount { get; set; }
-        public int AllowancePercentage { get; set; }
+        public double AllowancePercentage { get; set; }
         public double PayableAmount { get; set; }
         public DateTime CreatedOn { get; set; } = DateTime.Now;
         public List<InvoiceLine> InvoiceLine { get; set; }
